requirements.txt
SpeechRecognition==3.10.0
googletrans==3.1.0a0
gTTS==2.3.2
PyAudio==0.2.13
playsound==1.2.2
python-dotenv==1.0.0
Flask==2.3.3
Werkzeug==2.3.7 